// src/solana/enums/solana-signature-state.enum.ts
export enum SolanaSignatureStateEnum {
  NEW = 'NEW',
  PROCESSED = 'PROCESSED',
  IN_PROGRESS = 'IN_PROGRESS',
  NEW_LAST_SIGNATURE = 'NEW_LAST_SIGNATURE',
  PROCESSED_LAST_SIGNATURE = 'PROCESSED_LAST_SIGNATURE',
  IN_PROGRESS_LAST_SIGNATURE = 'IN_PROGRESS_LAST_SIGNATURE',
  NEW_FIRST_SIGNATURE = 'NEW_FIRST_SIGNATURE',
  PROCESSED_FIRST_SIGNATURE = 'PROCESSED_FIRST_SIGNATURE',
  IN_PROGRESS_FIRST_SIGNATURE = 'IN_PROGRESS_FIRST_SIGNATURE',
}